+++ requirements-cuda.txt
@@ -1,0 +1,2 @@
+torch>=2.1.0+cu118
+torchaudio>=2.1.0+cu118
