+++ requirements.txt
@@ -11,3 +11,18 @@ psutil
 beautifulsoup4
 whisper-cpp-pybind
 pydub
+numpy>=1.22.0
+soundfile>=0.12.1
+transformers==4.36.2
+TTS>=0.21.3
+requests>=2.31.0
+tqdm>=4.66.1
+importlib-metadata>=4.8.1
+packaging>=23.2
+pydantic>=1.10.13
+sounddevice>=0.4.6
+python-multipart>=0.0.6
+cutlet>=0.3.0
+unidic-lite>=1.0.8
+torch>=2.1.0+cu118
+torchaudio>=2.1.0+cu118
